+++ lib/src/widget/containers/containers.dart
@@ -1,3 +1,4 @@
+export 'box_fit.dart';
 export 'colored_box.dart';
 export 'constrained_box.dart';
 export 'container.dart';