+++ lib/src/widget/interact/gesture_detector.dart
@@ -1,4 +1,4 @@
-import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 import 'package:flutter/gestures.dart';
 
 extension WidgetOnGestureExtension on Widget {